requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
openai==1.26.0
pydantic==1.10.15
